prometheus: move from lib/ to imap/
We want lib/ to generally be things that work independently of an
installed and running Cyrus instance, and prometheus is not such.
